move (and rename) createTable function to more central file
